bg-remover-app/backend/requirements.txt
fastapi
uvicorn
python-multipart
rembg[cli]
pillow
onnxruntime
numpy
aiofiles
